Crate adml_dahlquist

Dahlquist Test Equation

A simple first-order ODE used to test numerical integrators:

dx/dt = -k * x

With k > 0, the analytical solution is: x(t) = x0 * exp(-k * t)

This is a fundamental test case for ODE solvers, useful for studying stability and accuracy of numerical integration methods.

Modules

ffi

Utilities related to FFI bindings.

fmi2

Dummy implementations for fmi2 functions

fmi3

Dummy implementations for fmi3 functions

ptr

Manually manage memory through raw pointers.

Structs

Dahlquist

Dahlquist test equation model

FmuInfo

Structure to store information about the FMU that will eventually be passed by some of the standard FMI functions. This structure can, optionally, be added to an struct that might need the information to work properly. The main intended use case is to be able to load resources based on a path that is dependent on the location of the FMU

PathBuf

An owned, mutable path (akin to String).

fmi2CallbackFunctions

Enums

FmiStatus

Enum that represents the states that an fmu model can return.

fmi2StatusKind

fmi2Type

fmi3DependencyKind

fmi3IntervalQualifier

Traits

FmuFunctions

Functions that must be implemented for any struct that uses this derive macro. The only required one is the do_step. The exit_initialization_mode has default behavior.

Functions

fmi3DeserializeFMUState

fmi3DoStep

fmi3EnterInitializationMode

fmi3ExitInitializationMode

fmi3FreeFMUState⚠

fmi3FreeInstance⚠

fmi3GetBoolean

fmi3GetFMUState⚠

fmi3GetFloat32

fmi3GetFloat64

fmi3GetInt8

fmi3GetInt16

fmi3GetInt32

fmi3GetInt64

fmi3GetString

fmi3GetUInt8

fmi3GetUInt16

fmi3GetUInt32

fmi3GetUInt64

fmi3GetVersion

fmi3InstantiateCoSimulation

fmi3Reset⚠

fmi3SerializeFMUState⚠

fmi3SerializedFMUStateSize

fmi3SetBoolean

fmi3SetFMUState⚠

fmi3SetFloat32

fmi3SetFloat64

fmi3SetInt8

fmi3SetInt16

fmi3SetInt32

fmi3SetInt64

fmi3SetString

fmi3SetUInt8

fmi3SetUInt16

fmi3SetUInt32

fmi3SetUInt64

Type Aliases

fmi2CallbackAllocateMemory

fmi2CallbackFreeMemory

fmi2CallbackLogger

fmi2StepFinished

fmi3ClockUpdateCallback

fmi3IntermediateUpdateCallback

fmi3LockPreemptionCallback

fmi3LogMessageCallback

fmi3UnlockPreemptionCallback

Derive Macros

Fmu

The function that implements the derive macro. It starts by passing the input token stream, with the help from the syn crate. In particular, the goal is to create different data structures with relevant information about the structure. Then, the parsed data structures are used to generate new functions for the structure that makes it support the FMI-standard.